pyproject.toml
[tool.poetry]
name = "clash of clan clock"
version = "0.1.0"
description = ""
authors = ["as6325400 <as6325400@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
line-bot-sdk = "^3.11.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
pymysql = "^1.1.1"
cryptography = "^42.0.8"
pytz = "^2024.1"
gunicorn = "^22.0.0"
uwsgi = "^2.0.26"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
